refactor(cpuhealth): split CPU signal table into per-signal files
table.go held every signal's declaration and the shared mark pairs, and was
too long for one file. Move each signal into its own file (signal_throttling,
signal_pressure, signal_steal, signal_saturation), each carrying the Marks var
that only it uses. A pure move: no signatures, names, or behaviour change; the
saturation family stays together because saturationRank and saturationFlags
order and flag both signals.
